# Conflicts:
#	.idea/workspace.xml

+++ app/src/main/java/tan/philip/nrf_ble/ScanListScreen/ScanResultsActivity.java
@@ -85,7 +85,7 @@ public class ScanResultsActivity extends AppCompatActivity {
                     mConnecting = false;
                     startGraphActivity((ArrayList<SignalSetting>) msg.getData().getSerializable("sigSettings"),
                             (BiometricsSet) msg.getData().getSerializable("bioSettings"),
-                            (int) msg.getData().getInt("notif f"));
+                            (float) msg.getData().getFloat("notif f"));
                     break;
                 case BLEHandlerService.MSG_UNRECOGNIZED_NUS_DEVICE:
                     new AlertDialog.Builder(ScanResultsActivity.this)
@@ -269,14 +269,14 @@ public class ScanResultsActivity extends AppCompatActivity {
 
 
     //Bluetooth methods
-    private void startGraphActivity(ArrayList<SignalSetting> signalSettings, BiometricsSet bioSettings, int notif_f) {
+    private void startGraphActivity(ArrayList<SignalSetting> signalSettings, BiometricsSet bioSettings, float notif_f) {
         Log.d(TAG, "Starting Graph Activity");
         Intent intent = new Intent(this, GraphActivity.class);
         Bundle extras = new Bundle();
         extras.putSerializable(GraphActivity.EXTRA_SIGNAL_SETTINGS_IDENTIFIER, signalSettings);
         extras.putSerializable(EXTRA_BIOMETRIC_SETTINGS_IDENTIFIER, bioSettings);
         extras.putString(EXTRA_BT_IDENTIFIER, getBluetoothIdentifier(mConnectingIndex)); //Probably not necessary, graph activity can ask for it from the service
-        extras.putInt(EXTRA_NOTIF_F_IDENTIFIER, notif_f);
+        extras.putFloat(EXTRA_NOTIF_F_IDENTIFIER, notif_f);
         intent.putExtras(extras);
 
         startActivity(intent);
